Adds cov_make_symmetric and enhances unit tests
Created a new aug_math module which includes the cov_make_symmetric function.
This function will make a covariance matrix symmetric if it is found to have
differences in off-diagonal elements. Updated the ric2eci function in the
cov_trans module to use cov_make_symmetric.

Updated and enhanced the cov_trans unit tests.

Added TODOs for other unit tests that need to be fully implemented.